requirements.txt
fastapi
uvicorn
opencv-python
numpy
mariadb
insightface
python-multipart
requests
onnxruntime-gpu
python-dotenv